category icon update and. decimal values smaller

+++ lib/widgets/add_transaction_form_income.dart
@@ -38,7 +38,7 @@ class _AddTransactionFormIncomeState extends State<AddTransactionFormIncome> {
       });
       final user = FirebaseAuth.instance.currentUser;
 
-      var amount = int.parse(amountEditController.text);
+      var amount = double.parse(amountEditController.text);
       DateTime date = DateTime.now();
 
       var id = uid.v4();
@@ -49,9 +49,9 @@ class _AddTransactionFormIncomeState extends State<AddTransactionFormIncome> {
           .doc(user!.uid)
           .get();
 
-      int remainingAmount = userDoc['remainingAmount'];
-      int totalCredit = userDoc['totalCredit'];
-      int totalDebit = userDoc['totalDebit'];
+      double remainingAmount = userDoc['remainingAmount'].toDouble();
+      double totalCredit = userDoc['totalCredit'].toDouble();
+      double totalDebit = userDoc['totalDebit'].toDouble();
 
       if (type == 'Credit') {
         remainingAmount += amount;
